optimize aminostring size

+++ rnalib/src/amino_string/counts.rs
@@ -1,13 +1,13 @@
 #[derive(Default, Clone, Copy)]
 pub struct Counts {
-	c: u32,
-	w: u32,
-	y: u32,
-	d: u32,
-	e: u32,
-	k: u32,
-	r: u32,
-	h: u32,
+	c: u16,
+	w: u16,
+	y: u16,
+	d: u16,
+	e: u16,
+	k: u16,
+	r: u16,
+	h: u16,
 }
 
 impl Counts {
@@ -25,35 +25,35 @@ impl Counts {
 		}
 	}
 
-	pub const fn get_c(&self) -> u32 {
+	pub const fn get_c(&self) -> u16 {
 		self.c
 	}
 
-	pub const fn get_w(&self) -> u32 {
+	pub const fn get_w(&self) -> u16 {
 		self.w
 	}
 
-	pub const fn get_y(&self) -> u32 {
+	pub const fn get_y(&self) -> u16 {
 		self.y
 	}
 
-	pub const fn get_d(&self) -> u32 {
+	pub const fn get_d(&self) -> u16 {
 		self.d
 	}
 
-	pub const fn get_e(&self) -> u32 {
+	pub const fn get_e(&self) -> u16 {
 		self.e
 	}
 
-	pub const fn get_k(&self) -> u32 {
+	pub const fn get_k(&self) -> u16 {
 		self.k
 	}
 
-	pub const fn get_r(&self) -> u32 {
+	pub const fn get_r(&self) -> u16 {
 		self.r
 	}
 
-	pub const fn get_h(&self) -> u32 {
+	pub const fn get_h(&self) -> u16 {
 		self.h
 	}
 }
